@@ -1,14 +1,13 @@
-﻿using MaxiCrush.MAUI.Views;
+﻿using MaxiCrush.MAUI.MVVM.Views;
 
-namespace MaxiCrush.MAUI
+namespace MaxiCrush.MAUI;
+
+public partial class App : Application
 {
-    public partial class App : Application
+    public App()
     {
-        public App()
-        {
-            InitializeComponent();
+        InitializeComponent();
 
-            MainPage = new AppShell();
-        }
+        MainPage = new AppShell();
     }
 }
